minify slicer by using thunk
