--- src/events/api/listEvents.ts
@@ -1,9 +1,0 @@
-import { DateTime } from 'luxon';
-import { getEvents } from './events';
-
-export const getListEvents = async () => {
-  return await getEvents({
-    event_end__gte: DateTime.local().toISODate(),
-    page_size: 10,
-  });
-};
