+++ src/events/slices/events.ts
@@ -1,10 +1,7 @@
-import { createAsyncThunk, createEntityAdapter, createSlice, SerializedError } from '@reduxjs/toolkit';
+import { createAsyncThunk, createEntityAdapter, createSlice, SerializedError, unwrapResult } from '@reduxjs/toolkit';
 import { State } from 'core/redux/Store';
-import { getCalendarEvents } from 'events/api/calendarEvents';
 import { getEvent, getEvents, IEventAPIParameters } from 'events/api/events';
-import { getImageEvents } from 'events/api/imageEvents';
-import { getListEvents } from 'events/api/listEvents';
-import { IEvent } from 'events/models/Event';
+import { IEvent, EventTypeEnum } from 'events/models/Event';
 import { DateTime } from 'luxon';
 
 const eventsAdapter = createEntityAdapter<IEvent>({
@@ -15,28 +12,58 @@ const eventsAdapter = createEntityAdapter<IEvent>({
 
 export const eventSelectors = eventsAdapter.getSelectors<State>((state) => state.events);
 
-export const fetchEvents = createAsyncThunk('events/fetchAll', async (options?: IEventAPIParameters) => {
+export const fetchEvents = createAsyncThunk('events/fetchMultiple', async (options?: IEventAPIParameters) => {
   const events = await getEvents(options);
   return events;
 });
+
 
 export const fetchEventById = createAsyncThunk('events/fetchById', async (eventId: number) => {
   const events = await getEvent(eventId);
   return events;
 });
 
-export const fetchEventList = createAsyncThunk('events/fetchList', async () => {
-  const events = await getListEvents();
-  return events;
+export const fetchEventList = createAsyncThunk('events/fetchList', async (_, { dispatch }) => {
+  const response = await dispatch(fetchEvents({
+    event_end__gte: DateTime.local().toISODate(),
+    page_size: 10,
+  }));
+  return unwrapResult(response);
 });
 
-export const fetchImageEvents = createAsyncThunk('events/fetchImageEvents', async () => {
-  const events = await getImageEvents();
-  return events;
+export const fetchImageEvents = createAsyncThunk('events/fetchImageEvents', async (_, { dispatch }) => {
+  const getTypeEvents = async (types: EventTypeEnum[]) => {
+    return await dispatch(fetchEvents({
+      event_end__gte: DateTime.local().toISODate(),
+      event_type: types,
+      page_size: 4,
+    }));
+  };
+  // Separate the three column fetches to be able to present some columns even if 1 type of events fails.
+  const left = getTypeEvents([EventTypeEnum.BEDPRES]);
+  const middle = getTypeEvents([EventTypeEnum.KURS]);
+  const right = getTypeEvents([
+    EventTypeEnum.SOSIALT,
+    EventTypeEnum.UTFLUKT,
+    EventTypeEnum.EKSKURSJON,
+    EventTypeEnum.ANNET,
+  ]);
+  const responses = await Promise.all([left, right, middle]);
+  responses.forEach(payloadCallback => unwrapResult(payloadCallback));
+  return responses;
 });
 
-export const fetchEventsByMonth = createAsyncThunk('events/fetchByMonth', async (month: DateTime) => {
-  const events = await getCalendarEvents(month);
+
+export const fetchEventsByMonth = createAsyncThunk('events/fetchByMonth', async (month: DateTime, { dispatch }) => {
+  const firstDayOfMonth = month.minus({ days: month.day - 1 });
+  const lastDayOfMonth = firstDayOfMonth.plus({ months: 1 }).minus({ days: 1 });
+
+  /** Set the query parameters of the fetch to the range, set page size large enough to get all */
+  const events = await dispatch(fetchEvents({
+    event_start__gte: firstDayOfMonth.toISODate(),
+    event_start__lte: lastDayOfMonth.toISODate(),
+    page_size: 60,
+  }));
   return events;
 });
 
@@ -78,40 +105,7 @@ export const eventsSlice = createSlice({
       builder.addCase(fetchEventById.rejected, (state, action) => {
         state.loading = 'idle';
         state.error = action.error;
-      }),
-      builder.addCase(fetchEventList.pending, (state) => {
-        state.loading = 'pending';
-      }),
-      builder.addCase(fetchEventList.fulfilled, (state, action) => {
-        state.loading = 'idle';
-        eventsAdapter.addMany(state, action.payload);
-      }),
-      builder.addCase(fetchEventList.rejected, (state, action) => {
-        state.loading = 'idle';
-        state.error = action.error;
-      }),
-      builder.addCase(fetchImageEvents.pending, (state) => {
-        state.loading = 'idle';
-      }),
-      builder.addCase(fetchImageEvents.fulfilled, (state, action) => {
-        state.loading = 'idle';
-        eventsAdapter.addMany(state, action.payload);
-      }),
-      builder.addCase(fetchImageEvents.rejected, (state, action) => {
-        state.loading = 'idle';
-        state.error = action.error;
-      }),
-      builder.addCase(fetchEventsByMonth.pending, (state) => {
-        state.loading = 'idle';
-      }),
-      builder.addCase(fetchEventsByMonth.fulfilled, (state, action) => {
-        state.loading = 'idle';
-        eventsAdapter.addMany(state, action.payload);
-      }),
-      builder.addCase(fetchEventsByMonth.rejected, (state, action) => {
-        state.loading = 'idle';
-        state.error = action.error;
-      });
+      })
   },
 });
